+++ src/Model/UserPasswordInterface.php
@@ -1,14 +1,12 @@
 <?php
 namespace Oka\ApiBundle\Model;
 
-use Symfony\Component\Security\Core\User\AdvancedUserInterface;
-
 /**
- * 
+ *
  * @author Cedrick Oka Baidai <okacedrick@gmail.com>
- * 
+ *
  */
-interface UserInterface extends AdvancedUserInterface
+interface UserPasswordInterface
 {
 	/**
 	 * @return string